+++ functions.php
@@ -4,7 +4,7 @@
  * Know the Code Genesis-powered child theme
  *
  * @package     KnowTheCode
- * @since       1.3.0
+ * @since       1.4.9
  * @author      hellofromTonya
  * @link        https://UpTechLabs.io
  * @license     GNU General Public License 2.0+
@@ -13,6 +13,3 @@ namespace KnowTheCode;
 
 require_once( 'lib/init.php' );
 require_once( 'lib/support/autoload.php' );
-
-// Launch Genesis
-include_once( get_template_directory() . '/lib/init.php' );