+++ lib/structure/footer.php
@@ -8,7 +8,7 @@
  * @link        https://UpTechLabs.io
  * @license     GNU General Public License 2.0+
  */
-namespace KnowTheCode;
+namespace KnowTheCode\Structure;
 
 /**
  * Unregister default navigation events.