+++ lib/structure/views/search/search-docx-entry.php
@@ -1,0 +1,3 @@
+<p <?php echo genesis_attr( 'entry' ); ?>>
+	<a href="<?php the_permalink(); ?>" rel="bookmark"><?php the_title(); ?></a>
+</p>