+++ lib/structure/views/search/search-header.php
@@ -1,0 +1,3 @@
+<div class="archive-description">
+	<h1 class="archive-title"><?php esc_html_e( $header_text ); ?></h1>
+</div>